% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacing.R
\name{korspacing}
\alias{korspacing}
\title{Spacing Korean Language}
\usage{
korspacing(text = NULL)
}
\arguments{
\item{text}{A character string to check spacing}
}
\value{
A spacing-checked string
}
\description{
Spacing Korean Language
}
\examples{
korspacing("아래와같은방식으로API를사용할수있으며,호출건수에대해서별도의제한은없으나,1회 호출에200글자로글자수를제한하고있다.")
}
